refactor: Concat device token datastore
