--- core/common/src/main/java/com/teamoffroad/core/common/data/datasource/DeviceTokenPreferencesDataSource.kt
@@ -1,8 +1,0 @@
-package com.teamoffroad.core.common.data.datasource
-
-import kotlinx.coroutines.flow.Flow
-
-interface DeviceTokenPreferencesDataSource {
-    val deviceToken: Flow<String>
-    suspend fun setDeviceToken(deviceToken: String)
-}
