+++ core/common/src/main/java/com/teamoffroad/core/common/data/datasource/TokenPreferencesDataSource.kt
@@ -5,8 +5,10 @@ import kotlinx.coroutines.flow.Flow
 interface TokenPreferencesDataSource {
     val accessToken: Flow<String>
     val refreshToken: Flow<String>
+    val deviceToken: Flow<String>
 
     suspend fun setAccessToken(accessToken: String)
     suspend fun setRefreshToken(refreshToken: String)
     suspend fun clearTokens()
+    suspend fun setDeviceToken(deviceToken: String)
 }
