+++ core/common/src/main/java/com/teamoffroad/core/common/data/di/DataModule.kt
@@ -2,9 +2,7 @@ package com.teamoffroad.core.common.data.di
 
 import com.teamoffroad.core.common.data.datasource.AutoSignInPreferencesDataSource
 import com.teamoffroad.core.common.data.datasource.DefaultAutoSignInPreferencesDataSource
-import com.teamoffroad.core.common.data.datasource.DefaultDeviceTokenPreferencesDataSource
 import com.teamoffroad.core.common.data.datasource.DefaultTokenPreferencesDataSource
-import com.teamoffroad.core.common.data.datasource.DeviceTokenPreferencesDataSource
 import com.teamoffroad.core.common.data.datasource.TokenPreferencesDataSource
 import dagger.Binds
 import dagger.Module
@@ -24,9 +22,4 @@ internal abstract class DataModule {
     abstract fun bindsAutoSignInLocalDataSource(
         dataSource: DefaultAutoSignInPreferencesDataSource,
     ): AutoSignInPreferencesDataSource
-
-    @Binds
-    abstract fun bindsDeviceTokenLocalDataSource(
-        dataSource: DefaultDeviceTokenPreferencesDataSource,
-    ): DeviceTokenPreferencesDataSource
 }
