+++ core/common/src/main/java/com/teamoffroad/core/common/data/di/RepositoryModule.kt
@@ -1,10 +1,8 @@
 package com.teamoffroad.core.common.data.di
 
 import com.teamoffroad.core.common.data.repository.AutoSignInRepositoryImpl
-import com.teamoffroad.core.common.data.repository.DeviceTokenRepositoryImpl
 import com.teamoffroad.core.common.data.repository.TokenRepositoryImpl
 import com.teamoffroad.core.common.domain.repository.AutoSignInRepository
-import com.teamoffroad.core.common.domain.repository.DeviceTokenRepository
 import com.teamoffroad.core.common.domain.repository.TokenRepository
 import dagger.Binds
 import dagger.Module
@@ -26,10 +24,4 @@ abstract class RepositoryModule {
     abstract fun bindAutoSignInRepository(
         authRepositoryImpl: AutoSignInRepositoryImpl,
     ): AutoSignInRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindDeviceTokenRepository(
-        deviceTokenRepositoryImpl: DeviceTokenRepositoryImpl
-    ): DeviceTokenRepository
 }
