--- core/common/src/main/java/com/teamoffroad/core/common/domain/repository/DeviceTokenRepository.kt
@@ -1,8 +1,0 @@
-package com.teamoffroad.core.common.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-
-interface DeviceTokenRepository {
-    val deviceToken: Flow<String>
-    suspend fun updateDeviceTokenEnabled(deviceToken: String)
-}
