+++ core/common/src/main/java/com/teamoffroad/core/common/domain/repository/TokenRepository.kt
@@ -15,4 +15,8 @@ interface TokenRepository {
     fun getRefreshToken(): Flow<String?>
 
     suspend fun clearTokens()
+
+    suspend fun updateDeviceTokenEnabled(deviceToken: String)
+
+    fun getDeviceToken():Flow<String>
 }
